Fix backtick and tilde on EU keyboards
On EU keyboards with vertical Enter key (both Czech and EU English) the key below Escape key outputs a Section sign (§) and Plus-minus sign (±) with Shift modifier. Nobody needs those signs.

Replace the key with a more sane values Grave accent (`) and Tilde (~). As it is on all the other keyboards. Programmers and system administrators expect them there because the Grave accent is often used in programming as backquote or backtick and tilde is used for home directories.