Update install.sh for Apple Silicon
Not sure if it's a Apple Silicon thing or a new macOS thing. But I got a new MBP last week and installed your amazing keyboard but unlike on my old MBP, it did not work. It did not add the keyboard to the Input Sources in System Preferences > Keyboard > Input Sources

I looked into it and that was the necessary change. 

I tested the script locally and it works.

Thank you for making such a handy tool! 🙏🏽